More combat fixes
Pvp armour is now lost on death, and the killer will get a coin value of
what the value of the item is.
Chaotics are now lost on death, and the killer will get 5M coins per
chaotic as the loot.
The delay on the korasi special attack has been fixed.
The dragon claw special has been fixed to have damage seperated into
each hit correctly.